requirements.txt
Django==4.1.1
dateutils==0.6.12
requests==2.28.1
